fixing toggle serialization
